+++ leap_ec/real_rep/ops.py
@@ -16,7 +16,6 @@ from leap_ec.ops import compute_expected_probability, iteriter_op
 ##############################
 # Function mutate_gaussian
 ##############################
-@curry
 def mutate_gaussian(std: float, expected_num_mutation: float = None,
                     hard_bounds: Tuple[float, float] = (-math.inf, math.inf)):
     """Mutate and return an individual with a real-valued representation.